qdslintplugin: warn about function blocks
Warn about bindings to functions outside of Connections or
ScriptActions.
Thats not 100% compatible with the QDS errors as we can't differentiate
between bindings to function and bindings to JS blocks, but I claim that
the end user doesn't care and that we can use the
ErrFunctionsNotSupportedInQmlUi category to cover all of these cases.

Fixes: QTBUG-129308
Change-Id: I1fbbb8b8a8e820ee64f6ad401992b1c956f01ae9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>