qmlls: don't mix up the in-memory and on-disk file
Fix QQmlJSImportVisitor to register the scope into its QQmlJSImporter.
This is important to avoid scope duplication.

Before this patch, the implicit directory import that
QQmlJSImportVisitor processes via QQmlJSImporter would create a
duplicate "lazy" scope that reads from the file on disk during
population.

With this patch, QQmlJSImporter will know the current scope and
therefore won't duplicate it during the implicit directory import.

Add a test to qmlls to make sure that it can lint enums correctly, even
if the enum declaration does not exist on the on-disk version.

Avoid warnings while resolving scopes by marking the currently to be
populated scope with the '$InProcess$' baseTypeName. It will be
populated (including with its real base type name) after the import
were processed.

Pick-to: 6.9 6.8
Fixes: QTBUG-134781
Change-Id: I05cc8f8cab9279ee07c4a48a1b467738354e11dd
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>