Disable edge-to-edge feature in qtquickview examples
After bumping the Android compile and target versions to 35,
disable edge-to-edge feature that arrives with Android 15, because it
will be easier for users to work with this for now.

Later we'll add support for edge-to-edge feature.

Task-number: QTBUG-134880
Pick-to: 6.8
Change-Id: I8d703c6c6b925a39da9ac8ea6a3d23eaa9b3c3a0
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 3d3f2dcb5ad3a0d57ae39383a9998196948c3d08)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>