Menu: adjust the insets according to the drop shadow
As it stood, the background delegate of a Menu would draw its
drop shadow on the outside of the delegate, and therefore also
on the outside of the Menu as a whole. This works fine for
Popup.Item, since an Item is allowed to draw out-of-bounds. But
for Popup.Window, this would fail.

Instead, the correct solution in this case is to draw the shadow
_inside_ the delegate, and instead make the whole menu bigger by
adjusting the insets instead, to take the shadow into account.

In order to be able to do this, we also need to teach StyleImage
to optionally draw the shadow inside the bounds.

Pick-to: 6.8
Change-Id: I306e511abca44f4f86ee18a16740f679cf987ac1
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit 39f6b9fb2a0b434e14a820a74de244163964659a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>