QQuickTableView: ensure focus transfer within the edit delegate works
Commit 48cb120c5f0bb5ce093b46bfe63751b4bd01caf9 broke focus transfer
within an edit delegate, since we would always close the editor upon
receiving a focus out event.

A typical use case is an edit delegate that contains an
editable ComboBox. In that case, focus will be transferred
within the delegate if the combo's textfield gains focus.

This patch will make sure that we only close the editor
if the edit item as a whole loses focus.

Task-number: QTBUG-134247
Pick-to: 6.9 6.8
Change-Id: If092ca91a96667eb2ee8ba271a2239086f5a2998
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>