quicklintplugin: check color names
Implement the ErrInvalidColor warning in qmllint via the quick plugin,
where literal strings assigned to colors are checked for correct color
names.

Add an assert in the constructor to make sure that the colors are
sorted. We use a QStringList instead of a constexpr std::array
to store the const list of color names because we need a QStringList
for the "Did you mean" suggestions anyway.

Task-number: QTBUG-129307
Change-Id: Ifbe0ce7f5158be7dae524889944f351184cc8559
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>