qmllint: register missing-enum-entry category
Amends d876ce343d5875ec132aa1b878970f03cbae353d and add a description
and a settingname to the missing-enum-entry category, so that it can get
registered into qmllint.

Also fix the test to expect a warning instead of an info message. Info
messages are selected by default for unregistered categories.

Task-number: QTBUG-135020
Change-Id: I1fe3684bb12c198633bb86e1cb96a390f352739b
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>