qmlformat: customizable semicolon
Add semicolon option to qmlformat. While --Always always appends
semicolons to the JS statements, --essential removes the semicolons
unless it is not safe to rely on ASI once semicolons are removed.

Change the way EmptyStatements are handled. Prior to this
commit, semicolons following if, for, foreach, while statements
without bodies were added to the new line with some indentation.
Make the semicolon following no-body items stick to the right
paranthesis. If there is a chain of empty statements, write out a
single one.

[ChangeLog][qmlformat] New option semicolon-rule is added and
EmptyStatement formatting behavior has changed.

* Added --semicolon-rule option with modes:
  - --semicolon-rule=always: Always appends semicolons to JS statements.
  - --semicolon-rule=essential: Removes semicolons unless ASI makes it unsafe.

* Changed handling of EmptyStatements:
  - Semicolons after control structures without a body (e.g., if, for, while)
    now appear directly after the closing parenthesis, instead of on a new line.
  - Consecutive empty statements are collapsed into a single semicolon.

Fixes: QTBUG-107152
Change-Id: Ic95047a1f0077937d4c1f01328d77a3e6a4f22d6
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>