Close all other Menus when showing non-sub-menu
This fixes the linked bug, but also generally makes sense,
as there shouldn't be a use case where two menus are open at once.

Adjust the documentation to reflect the current state of
context menus and use best practices.

[ChangeLog][Controls][Important Behavior Changes] All non-sub-menus
are now closed when a menu is opened. This is to fix an issue
(QTBUG-134903) where duplicate menus are shown when opening a custom
non-ContextMenu on a text editing control like TextField or TextArea.
This means that rather than two context menus incorrectly being opened,
an extra one will briefly be visible before immediately being closed.
For information on how to avoid this, please see the "Context menus"
section of Menu's documentation.

Fixes: QTBUG-134903
Pick-to: 6.9
Change-Id: I158295b72176935c2dc8072a72a4df66678be824
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>