qmllint: enhance var-used-before-declaration warning
It doesn't really make sense to print out the location of the variable
declaration, especially in the language server. Instead, emit two
warnings: one at the declaration and one at the usage. This makes it
easier for users to spot the location of the declaration.

Add a test to make sure that both warnings can be disabled with one
disabling directive.

Task-number: QTBUG-129307
Change-Id: Icb485ee37f016f18395c3016672e397c4f5e6024
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>