tst_qmllint: add shorthand to enable categories
Add a CallOption struct that allows to control how qmllint is invoked,
and an callQmllint() overload that accepts this CallOption struct.
Currently it only allows to enable disabled categories. This is useful
to test warnings that are disabled by default, for example the
"function-used-before-declaration" opt-in category that will be added in
a later commit.

Task-number: QTBUG-129307
Change-Id: I82b2201e1389dd1f95364109e159eaa5f5b87115
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>