qmllint: suppress secondary messages with main message
Allow warnings to be emitted in multiple parts, as a preparation step to
have two warnings for "var-used-before-declaration": one of the usage
and one for the declaration.

Added an extra parameter to log that overrides the line number used to
determine whether the current warning is ignored or not. This assumes
that both main and secondary messages both use the same warning
category.

Task-number: QTBUG-129307
Change-Id: I82ed01d417070b13e3c91ea82218e0d858e10d24
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>