QtQml: Do not store property indices in aliases
We shouldn't store property indices in the compilation units. That makes
the system brittle and vulnerable to changes in unrelated code. Rather,
we should always resolve aliases from names at run time. We can store
the alias target property indices in the property caches. That's where
they belong.

This means we need to add the alias properties to the property cache
right when we resolve the aliases since we can't communicate the
property indices back anymore.

Task-number: QTBUG-135286
Change-Id: I1d25f66ee06cff3fdf958f766c62cb99a30f6720
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>