doc: Make transform snippets runnable
It's not clear at this point where qt.png was supposed to come from, but
the modern Qt logo seems to work well enough. In the case of Item with
two transforms, perhaps this was not tested before; that's the trouble
with inline qml code. It's a little easier to visualize the effect with
an Image of which you can see more than one edge (although parts of
it remain hidden now too).

Pick-to: 6.8
Change-Id: I64c685d4dc9568080b0ddb4921d987b07eb17c41
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 36843b1bf7a12f9a54b6e484722c6bafb88df9e9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>