vectorimage: Fix identical gradient stops in qml generator
In SVG, two subsequent identical gradient stops means a hard
transition to the color of the second at that position. However,
this is not the case in Qt and because of the way we sort the
gradient stops, the order of the two identical stops will be
wrong.

This is addressed by adding an epsilon to identical stops in
Qt Svg, but the epsilon is FLT_EPSILON, which is around 1e-7.
The default output precision for doubles in Qt is 6, so when
storing the numbers to file, the epsilon would be lost, the
two positions would be considered identical and the reordering
would happen.

This caused a visual error when running svgtoqml on the
paint-grad-16-t.svg test.

Pick-to: 6.8 6.9
Change-Id: I857958512c3de7dc843a705893b4c2d7d2a36fd4
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>