qqmljslinter: use early exit and remove huge lambda
Clean up QQmlJSLinter::lintFile a bit. Add an early exit to decrease the
indent for the qml file linting code.

Remove the huge "check" lambda: instead, create an optional
QQmlJSFileResourceMapper to support both use cases of linting with or
without file resources. This makes the lambda useless, so remove the
calls of the lambda and move the lambda code directly at the caller
site.

Change-Id: Ib268f3e40d945f41710675f2bdb534aa62a599f0
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>