doc: Make the overlay less confusing by documenting what's going on
Event handling in QtQuick is already quite complicated. It doesn't help
that the overlay increases this complexity by filtering events in two
separate steps.

We'll hopefully be able to refactor some of this code in the future, but
before we can do that, we need to understand how this works currently.

Change-Id: I95bdd76d23e41a2293fa7dfdd401af980fc0be51
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>