CMake: Force override while code signing qmltestrunner
Some clang toolchains use implicit adhoc codesigning when compiling
and linking executables. This causes an error for the qmltestrunner
codesign call, saying that the executable is already signed.

Force the codesign command to override the existing signature in that
case. This is safe, because we build the executable from scratch.

Amends 1f7b87cd3d0dff022b3c072016c6257cc5d9e93a

Pick-to: 6.8 6.9
Fixes: QTBUG-135279
Change-Id: I0e0a989eb018c75ad1463700782bef9c7b999fa0
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>