Add API to VectorImage for controlling animations
A lot of animations will just be a single time line that you
is intended to be controlled from the outside (and formats like Lottie
does not support anything beyond this), so we need some API to
match the general Animation API so that you can restart, pause/resume
and change the number of times an animation loops.

We do this by adding a group of properties called "animations" that
can be used to control all the animations in the document at once.

We generate this both in the QML file from *toqml and also in
VectorImage itself.

[ChangeLog][QtQuickVectorImage] Added some API to the VectorImage
for looping, stopping, pausing and resuming animations.

Fixes: QTBUG-135265
Change-Id: Id372c00110d165d02db357ce77eb1dec504cffb8
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>