Respect target DPR when drawing text objects in QQuickTextNodeEngine
Instead of passing the DPR through as function arguments to each addFoo
function we set the DPR on the QSGInternalTextNode and QQuickTextNodeEngine.
We could have solved this by pulling the DPR from the QSGRootNode's
renderer, but there might be more than one of them, and we're missing
a setDevicePixelRatio in QSGAbstractRenderer (it's only exposed one
level above, in QSGRenderer).

Pick-to: 6.9
Fixes: QTBUG-127913
Change-Id: I48081d441259f0713cdc5f784eede6777b5fb601
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>