qqmljscontextproperties: "grep" for setContextProperty calls
Add code to Grep the project folder for "setContextProperty" calls, so that
a later commit can warn about usages of those context properties.

Use grep to find files with setContextProperty calls, and then run a
QRegularExpression on the files to extract the names and locations of
the context properties.

Add a fallback in case there is no grep on the system, and don't run
grep on windows. Also avoid unused functions errors on windows where
grep is not run so its output also does not need to be parsed.

Task-number: QTBUG-128232
Change-Id: I15ceb2bbeb6ae6ac83f4bfa85b8a44d5a897d0a7
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>