Introduce a TreeModel QML type
This is a QML type that can be used as a model for the TreeView.

The model data can be defined in the QML file. After the data
has been assigned for the first time, the columns and roles are
fixed for the lifetime of the model (just like in the case of
TableModel QML type).

The API is designed to be similar to the API of TableModel.

[ChangeLog][Qt labs QML Models] Added a new TreeModel QML type
that allows the definition of a tree structure in the QML file
and works with TreeView.

Task-number: QTBUG-130571
Change-Id: Ib73d1b79de1ab6505b81126bc676935ae975c47e
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>