QQmlPropertyPrivate::write: Move list handling in helper function
The main method is rather unwieldy, so dispatch the code to handle list
properties to a helper function.

We also return now early; this does not make any difference, because
write will just unconditionally return true after the if-else cascade.
Returning the return value of writeProperty is also not a change, as it
unconditionally returns true, too.

Change-Id: Ib2c47052f75dc4098900107c67448d05ca8f48e5
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>