tst_qmllint: clean callQmllint arguments
callQmllint requires so many parameters that it is getting quite
unwieldy: move its arguments into a struct, so that its easier to
distinguish between actually relevant arguments
and unchanged defaulted arguments.

Introduce CallQmllintOptions that contain information on how to call
qmllint, and CallQmllintChecks that contain information on what tests
callQmllint does (currently it can checks autofixes and the return
value of qmllint).

Remove the QJsonArray* argument and always return the array.
Keep the old behavior of not populating the array when a test failure
happened to avoid breaking tests with QEXPECT_FAIL().

Change-Id: I9dffbea0dee5033df748beec72abf00e5a141cd5
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>