Use platform theme's palette in macOS style
Unlike most other styles the macOS style didn't initialize QQuickTheme
with a style specific palette (either directly or via a custom theme
class).

Instead, we would indirectly pick up the platform theme palette when
QQuickStylePlugin::createTheme called QQuickStylePrivate::readPalette
to read palette settings, as we then default-constructed a QPalette,
which picks up its colors from QGuiApplicationPrivate::app_pal, which
in turn is based on platform theme palette, via QGuiApp's basePalette.

However QQuickStylePlugin::createTheme only initialized the System
palette of QQuickTheme, which meant that we failed to use any role
specific palettes that the platform theme provided.

In addition, because we only initialized the palette once, when
creating the theme, we failed to pick up any changes to the platform
theme's palette.

To fix this we simply tell the QQuickTheme to prefer
the platform theme for its palette, which means any query to the theme
will go through the platform theme. This also means we don't need to
hook into QtQuickControls2StylePlugin::updateTheme() to re-initialize
the palette.

Technically we do have the same problem with the theme's fonts, but for
that we don't have a way to instruct the theme to use the platform theme.

Pick-to: 6.9 6.8 6.5
Change-Id: I56eb278a80b184397114282839e958d61bd0028e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>