macOS: Port from deprecated NSAppearance.currentAppearance
When the system appearance changes, it will be reflected through
NSApp.effectiveAppearance, but NSAppearance.currentAppearance
and NSAppearance.currentDrawingAppearance will remain as is for
some reason (perhaps as a compatibility for apps that were not
ready to handle dark mode).

The original way to deal with this was to explicitly set the
NSAppearance.currentAppearance to the new effective appearance,
which would take effect for the thread from that point on. But
this API has been deprecated, most likely because overriding it
globally could mess up views or logic that was not prepared for
dark mode).

The replacement API, NSAppearance performAsCurrentDrawingAppearance
is given a block, and will only override NSAppearance.currentAppearance
during the block, which isolates it from other fragile components.

We now use a helper from QtGui that takes care of wrapping our
style drawing in performAsCurrentDrawingAppearance.

Task-number: QTBUG-135789
Change-Id: I0eb5dfc05734cf75181e6a3ed8cf2d23185f425f
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>