Set the current directory earlier in tst_QQuickFileDialogImpl
The temporary directory is set as the current directory at the end of
iniTestCase. iniTestCase() returns earlier if the temporary directory
is not empty and this code is never reached.

Set the current directory directory as the temporary directory earlier
in tst_QQuickFileDialogImpl::initTestCase() to ensure that it is
always set properly.

Fix the failing test
tst_QQuickFileDialogImpl::changeFolderViaDoubleClick() on Ubuntu 24

Pick-to: 6.9 6.8 6.5
Change-Id: I9e124280e69c67056e0cc9ce15a1a45b57221cb2
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>