Support replace animations
SVG has a feature where a set of animations on the transform
can be set to either replace or append to previous animations,
but only while the animation set is active.

In order to support this, we introduce a helper class called
TransformGroup which can contain a set of parallel animations
and have them override the others.

Change-Id: I28be7d120527b1f2d1e3dbf5c5de94094c766363
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>