qmllint: Fix required property warning with owner mixup
Because we only search for the name of the property in the
scopesToSearch and not which property that name actually references, we
confuse a property of the same name in a neighboring scope for the
actual required one.

Include the property's owning scope when searching through the scopes.

Amends daf57e29de918b7b4be7bb0d469db0c51d41bb07

Fixes: QTBUG-136058
Pick-to: 6.9 6.8
Change-Id: I998901fd0840270dd2048e7257d6eaca556b513d
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>