qmltyperegistrar: Do not filter members by revision for qmltypes
We need information about all members in the qmltypes file, even if the
revisions are nonsensical. This is because we need to get the method
indices right when trying to call methods.

We don't want to actually register the extra versions, but the code
handling registration is different from the code producing qmltypes
already.

Amends commit c0f4442ae0d83e63fc90f1d5058081355f83b701

Pick-to: 6.9
Fixes: QTBUG-136248
Change-Id: Ibd41fcd3a3d7b9d79e1d1715746dd90702ade518
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>