qquickpopup: Capture activeFocusItem before reparenting popup
Previously, prepareEnterTransition would reparent the popupItem before
saving window->activeFocusItem(), which could shift focus into the
popup and cause the saved focus to fall back to the window contentItem.

In particular, during reparenting focus may be set if user code calls
forceActiveFocus, for example, in the onVisibleChanged handler.

The activeFocusItem is now captured before reparenting and stored in
lastActiveFocusItem.

Pick-to: 6.9
Change-Id: I6f4691e871661b6ba66b43ed720dc771a65ae3ed
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>