QQmlComponent::loadFromModule: Fix crash with dynamic metaobjects
If an object has a dynamic meta-object (e.g. because it has been
extended with additional enums), we can no longer create a property
cache.
This could have caused a null-poninter dereference; avoid this by
checking for nullptr.

We currently don't do any checks at all in that case; finding a better
solution is tracked in QTBUG-136560.

Pick-to: 6.9 6.8
Change-Id: I678cf9908d5bbec50e133a462f7f4c813dca44dd
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>