Menu: force contentItem ListView layout before starting enter transition
Enter transitions may want to animate the Menu's height based on its
implicitHeight. The Menu's implicitHeight is typically based on the
ListView's contentHeight, among other things. The docs for ListView's
forceLayout function say: "Responding to changes in the model is
usually batched to happen only once per frame." As e.g.
NumberAnimation's from and to values are set before any polishes
happen, any re-evaluation of their bindings happen too late, and the
starting height can be out-dated when menu items are added after
component completion (QQuickItemView::componentComplete does a layout,
so items declared as children aren't affected by this).

To account for this, this patch forces a layout before the transition
starts, if necessary.

Fixes: QTBUG-136256
Pick-to: 6.5 6.8 6.9
Change-Id: I1cc912347b774369b3542f0d32c654ae5615b97a
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>