tst_qmllint: Clean up sorted functions
Since we recently sorted some data test functions and the git history
already is compromised, we can take the opportunity to clean those up.

This unifies the style of all test entries and removes a lot of
boilerplate. This should make the tests easier to read and navigate.

There were 517 tests executed by qmllint both before and after this
change.

Pick-to: 6.9 6.8 6.5
Change-Id: Ie394c5b6e967377903da0138ac7ee37ec9dc6765
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>