QtQml: Do not prematurely delete composite metatypes
We may store them in various places, for example as argument types to
methods. Rather, keep a cache of composite metatypes and recycle them as
necessary. These days, each type has a unique URL and we can use those
URLs to recognize them. If a type is dropped and later re-loaded we can
simply change the name of the metatype in place. It's the only actually
unique piece of data about composite metatypes.

As a drive-by simplify the creation of names for inline component types
a bit and move the names into place.

Pick-to: 6.9 6.8
Fixes: QTBUG-135286
Change-Id: I01728a1e5f40175072d8a27bb1b0dee5929d31c2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>