FluentWinUI3: Improve CheckBox when using high contrast themes
The CheckBox indicator uses a ColorImage to blend the pixels of a
png image file together with a QColor, in order to determine the colors
that the indicator should use in any given state (e.g, checked,
unchecked, hovered, etc).

Unfortunately, blending the ColorImage with a color has its limitations
when enabling a high contrast theme on Windows 11, such as when the
checkbox is in the pressed and unchecked state, where the color image
background will typically be mostly black, and this make the color
bound on the color property much darker, or in this case, completely
black.

To get around this issue, we now have a Rectangle on top of the
ColorImage, which is normally transparent, except for when high contrast
is enabled. In that case, it will be rendered on top, and use the exact
desired colors.

Task-number: QTBUG-129088
Change-Id: Ic65218d3759495611772f010c86a85109faaaadd
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 91690dbaa752d6230d050a5d78cbfd89fc26ec05)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>